:sparkles: [DisaffiliateBeneficiary] Move route into UserController, use User subject
