+++ src/ControllerV2/UserController.php
@@ -2,6 +2,7 @@
 
 namespace App\ControllerV2;
 
+use App\Entity\Centre;
 use App\Entity\User;
 use App\FormV2\ChangePasswordFormType;
 use App\FormV2\UserAffiliation\AffiliateUserType;
@@ -9,7 +10,9 @@ use App\FormV2\UserAffiliation\Model\AffiliateUserModel;
 use App\FormV2\UserSettingsType;
 use App\ManagerV2\RelayManager;
 use App\ManagerV2\UserManager;
+use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\EntityManagerInterface;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use Symfony\Component\Form\FormError;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -108,5 +111,39 @@ class UserController extends AbstractController
         }
 
         return $this->render('v2/user/invite.html.twig', ['form' => $form]);
+    }
+
+    #[Route(
+        path: '/{id}/disaffiliate/choose-relay',
+        name: 'disaffiliate_relay_choice',
+        requirements: ['id' => '\d+'],
+        methods: ['GET'],
+    )]
+    #[IsGranted('UPDATE', 'user')]
+    public function disaffiliateChooseRelay(User $user): Response
+    {
+        return $this->render('v2/user_affiliation/beneficiary/disaffiliate_beneficiary.html.twig', [
+            'user' => $user,
+            'relays' => $this->getProfessional()?->getManageableRelays($user->getSubject()) ?: new ArrayCollection([]),
+        ]);
+    }
+
+    #[Route(
+        path: '/{id}/relay/{relayId}/disaffiliate',
+        name: 'disaffiliate_user',
+        requirements: ['id' => '\d+'],
+        methods: ['GET'],
+        condition: 'request.isXmlHttpRequest()',
+    )]
+    #[ParamConverter('relay', class: 'App\Entity\Centre', options: ['id' => 'relayId'])]
+    #[IsGranted('UPDATE', 'user')]
+    public function disaffiliateFromRelay(
+        User $user,
+        Centre $relay,
+        RelayManager $manager,
+    ): Response {
+        $manager->removeUserFromRelay($user, $relay);
+
+        return $this->json($user);
     }
 }
