+++ templates/v2/professional/_beneficiaries_list_row.html.twig
@@ -21,7 +21,7 @@
     <td class="border-0 h-100 text-center align-middle" {{ stimulus_controller('dropdown-menu') }}>
         {{ include('v2/professional/_action_dropdown.html.twig', {
             removeFromRelay: {
-                path: path('disaffiliate_beneficiary_relay_choice', {'id': beneficiary.id}),
+                path: path('disaffiliate_relay_choice', {'id': beneficiary.user.id}),
                 message: 'remove_from_relay_confirm'|trans({'%beneficiairy%': beneficiary.user.fullName})
             },
             resetPasswordPath: path('reset_password_beneficiary', {'id': beneficiary.id}),
