+++ templates/v2/user_affiliation/beneficiary/disaffiliate_beneficiary.html.twig
@@ -1,13 +1,13 @@
 {% extends 'v2/management_layout.html.twig' %}
 
 {% block title %}
-    {{ 'disaffiliate_beneficiary_title_part_1'|trans }}
+    {{ 'disaffiliate_user_title_part_1'|trans }}
     <span class="text-blue">
-        {{ 'disaffiliate_beneficiary_title_part_2'|trans }}
+        {{ 'disaffiliate_user_title_part_2'|trans }}
     </span>
 {% endblock %}
 {% block subTitleIcon %} <i class="fas fa-user-times fa-lg me-1"></i> {% endblock %}
-{% block subTitle %}{{ 'disaffiliate_beneficiary_subtitle'|trans|upper }} {{ beneficiary.user.fullName|upper }} {% endblock %}
+{% block subTitle %}{{ 'disaffiliate_beneficiary_subtitle'|trans|upper }} {{ user.fullName|upper }} {% endblock %}
 
 {% block content %}
     <h5 class="text-center my-4 px-3 px-md-5">{{ 'select_relays_for_disaffiliation'|trans }} : </h5>
